devmem: remove the use of long type for python 3 conformance
The type long is not available anymore in python 3. The usual fix is
to simply use the int type.
We remove the use of long type in any case (even with python 2) because
the address passed to devmem is currently 32 bit word that fits in an int type.

64 bits word with python 2 kind of usage is unlikely to happen on a Microchip
MPU as the  move to python 3 is nearly completed for everyone already.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>